+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 - Fix missing property errors
 - Ignore destinations without mount point (until a better solution is found)
+- Scroll to selected item when using arrow keys
 
 ## 1.2.0 - 2023 May 19
 - Support macOS 13